Output minimal workflow
